+++ src/Arrays/param.hh
@@ -77,7 +77,7 @@ struct GALFIT_PAR : GALMOD_PAR {
     string SIDE       = "B";      ///< Approaching(A), Receding(R), Both(B), Single(S)
     bool   TWOSTAGE   = true;     ///< Whether fitting a second model after regularization.
     string REGTYPE    = "auto";   ///< Regularazation type for second stage
-    int    BWEIGHT    = 1;        ///< Power of the weighting function for Blank pixels.
+    float  BWEIGHT    = 1;        ///< Power of the weighting function for Blank pixels.
     int    STARTRAD   = 0;        ///< Starting radius
     bool   flagERRORS = false;    ///< Whether estimating errors.
     float  DISTANCE   = -1;       ///< Distance of the galaxy to convert arcs to kpc.
